% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PolyaTree.R
\name{dPTdist}
\alias{dPTdist}
\title{Density function for a distribution function}
\usage{
dPTdist(X, PT.model, dist.sample, ddist, pdist, log = FALSE)
}
\arguments{
\item{X}{values to be checked.}

\item{PT.model}{Polya tree model, either from PT.prior or PT.posterior.}

\item{dist.sample}{probability function previously drawn from the PT.}

\item{ddist}{centering density function from the prior.}

\item{pdist}{centering distribution function from the prior.}

\item{log}{return the logarithm of the density, default is FALSE.}
}
\description{
Density function for a distribution function drawn from the PT process.
}
\examples{
#Drawing from data 
set.seed(42);
Y <- rnorm(100);

#Posterior based on prior model
post <- PT.posterior(Y = Y, qdist = qnorm, layers = 3);

#Sample 1 distribution from the PT
dist.sample <- rPT(n = 1, PT.model = post, seed = 42)[,,1];

#Sample 5 observations from each distribution
print(
    dPTdist(X = c(-1,0,1), PT.model = post, dist.sample = dist.sample,
    ddist = dnorm, pdist = pnorm, log = FALSE));
}
