% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PolyaTree.R
\name{pPTdist}
\alias{pPTdist}
\title{Distribution function}
\usage{
pPTdist(X, PT.model, dist.sample, pdist, log = FALSE)
}
\description{
Distribution function for a distribution function drawn from the PT process.
}
\examples{
#Drawing from data 
set.seed(42);
Y <- rnorm(100);

#Posterior based on prior model
post <- PT.posterior(Y = Y, qdist = qnorm, layers = 3);

#Sample 1 distribution from the PT
dist.sample <- rPT(n = 1, PT.model = post, seed = 42)[,,1];

#Sample 5 observations from each distribution
print(
    pPTdist(X = c(-1,0,1), PT.model = post, dist.sample = dist.sample,
    pdist = pnorm, log = FALSE)
    );
}
