% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PolyaTree.R
\name{PT.posterior}
\alias{PT.posterior}
\title{Polya tree posterior}
\usage{
PT.posterior(
  Y,
  qdist = NULL,
  layers = ceiling(log(length(Y), base = 2)),
  vartype = "continuous",
  v = 1,
  PT.model = NULL,
  cores = 1
)
}
\arguments{
\item{Y}{dataset, must be a vector.}

\item{qdist}{quantile function of the centering distribution, default is NULL.}

\item{layers}{number of layers of the prior, default is the log of the sample size in base 2.}

\item{vartype}{variable type, default is 'continuous'.}

\item{v}{hyperparameter related to how close the prior is to the centering distribution, default is 1.}

\item{PT.model}{Polya tree prior to be updated, default is NULL.}

\item{cores}{number of cores for parallel processing, default is 1.}
}
\description{
Posterior of the Polya tree process based on PT.prior and data
}
\examples{
#Drawing from data 
set.seed(42);
Y <- rnorm(100);

#Posterior based on prior model
prior <- PT.prior(qdist = qnorm, layers = 3);
print(PT.posterior(Y = Y, PT.model = prior));

#Posterior without prior model
print(PT.posterior(Y = Y, qdist = qnorm, layers = 3));
}
