% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PolyaTree.R
\name{PT.prior}
\alias{PT.prior}
\title{Polya tree prior}
\usage{
PT.prior(qdist, layers, vartype = "continuous", v = 1)
}
\description{
Specifies the partially specified Polya tree prior
}
\examples{
print(PT.prior(qdist = qnorm, layers = 3));
}
