% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PolyaTree.R
\name{qPTdist}
\alias{qPTdist}
\title{Quantile function}
\usage{
qPTdist(p, PT.model, dist.sample, pdist, qdist)
}
\arguments{
\item{p}{probabilities to be checked.}

\item{PT.model}{Polya tree model, either from PT.prior or PT.posterior.}

\item{dist.sample}{probability function previously drawn from the PT.}

\item{pdist}{centering distribution function from the prior.}

\item{qdist}{centering quantile function from the prior.}
}
\description{
Quantile function for a distribution function drawn from the PT process.
}
\examples{
#Drawing from data 
set.seed(42);
Y <- rnorm(100);

#Posterior based on prior model
post <- PT.posterior(Y = Y, qdist = qnorm, layers = 3);

#Sample 1 distribution from the PT
dist.sample <- rPT(n = 1, PT.model = post, seed = 42)[,,1];

#Sample 5 observations from each distribution
print(
    qPTdist(p = c(.2,.5,.8), PT.model = post, dist.sample = dist.sample,
    pdist = pnorm, qdist = qnorm));
}
