% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PolyaTree.R
\name{rPT}
\alias{rPT}
\title{Draw from the Polya tree process}
\usage{
rPT(n, PT.model, seed = NULL)
}
\arguments{
\item{n}{sample size.}

\item{PT.model}{Polya tree model, either from PT.prior or PT.posterior.}

\item{seed}{seed for sampling, default is NULL.}
}
\description{
Draws a probability distribution function based from the Polya tree process. Last column is the log probability of sampling an observation from that interval.
}
\examples{
#Drawing from data 
set.seed(42);
Y <- rnorm(100);

#Posterior based on prior model
post <- PT.posterior(Y = Y, qdist = qnorm, layers = 3);

#Sample 10 distributions from the PT
print(rPT(n = 10, PT.model = post, seed = 42));
}
