% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PolyaTree.R
\name{rPTdist}
\alias{rPTdist}
\title{Draw observations from the probability distributions of the PT}
\usage{
rPTdist(n, PT.model, dist.samples, pdist, qdist, seed = NULL)
}
\description{
Draws observations from probability distribution functions generated from the Polya tree process.
}
\examples{
#Drawing from data 
set.seed(42);
Y <- rnorm(100);

#Posterior based on prior model
post <- PT.posterior(Y = Y, qdist = qnorm, layers = 3);

#Sample 10 distributions from the PT
dist.samples <- rPT(n = 10, PT.model = post, seed = 42);

#Sample 5 observations from each distribution
print(rPTdist(n = 5, PT.model = post, dist.samples = dist.samples, pdist = pnorm, qdist = qnorm, seed = 42));
}
