% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PolyaTree.R
\name{PT.posterior}
\alias{PT.posterior}
\title{Polya tree posterior}
\usage{
PT.posterior(
  Y,
  qdist = NULL,
  layers = ceiling(log(length(Y), base = 2)),
  vartype = "continuous",
  v = 1,
  PT.model = NULL,
  cores = 1
)
}
\description{
Posterior of the Polya tree process based on PT.prior and data
}
\examples{
#Drawing from data 
set.seed(42);
Y <- rnorm(100);

#Posterior based on prior model
prior <- PT.prior(qdist = qnorm, layers = 3);
print(PT.posterior(Y = Y, PT.model = prior));

#Posterior without prior model
print(PT.posterior(Y = Y, qdist = qnorm, layers = 3));
}
