% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PolyaTree.R
\name{theta.part}
\alias{theta.part}
\title{Theta partition at each layer.}
\usage{
theta.part(layer)
}
\arguments{
\item{layer}{number of layers.}
}
\description{
Sample space partitions based on the quantile function for each layer. Except for infinities, intervals are closed on the left and open on the right.
}
